@@ -13,7 +13,6 @@ import (
 	"github.com/metal-stack/metal-lib/pkg/pointer"
 	"github.com/metal-stack/metal-lib/pkg/testcommon"
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
 	"google.golang.org/protobuf/runtime/protoimpl"
 
@@ -28,7 +27,7 @@ func TestMain(m *testing.M) {
 	defer func() {
 		os.Exit(code)
 	}()
-	log = slog.Default()
+	log = slog.New(slog.NewJSONHandler(os.Stdout, &slog.HandlerOptions{Level: slog.LevelDebug}))
 	code = m.Run()
 }
 
@@ -112,7 +111,7 @@ func TestDeleteTenant(t *testing.T) {
 	var paging *v1.Paging
 
 	storageMock.On("Delete", ctx, t3.Meta.Id).Return(nil)
-	memberStorageMock.On("Find", ctx, tfilter, paging).Return([]*v1.TenantMember{
+	memberStorageMock.On("Find", ctx, paging, tfilter).Return([]*v1.TenantMember{
 		{
 			Meta: &v1.Meta{
 				Id: "t3",
@@ -121,7 +120,7 @@ func TestDeleteTenant(t *testing.T) {
 			MemberId: t3.Meta.Id,
 		},
 	}, nil, nil)
-	memberStorageMock.On("Find", ctx, mfilter, paging).Return([]*v1.TenantMember{
+	memberStorageMock.On("Find", ctx, paging, mfilter).Return([]*v1.TenantMember{
 		{
 			Meta: &v1.Meta{
 				Id: "t3",
@@ -160,47 +159,250 @@ func TestGetTenant(t *testing.T) {
 	assert.Equal(t, tgr.Id, resp.GetTenant().GetMeta().GetId())
 }
 
-func TestFindTenantByID(t *testing.T) {
-	storageMock := &mocks.Storage[*v1.Tenant]{}
-	ts := &tenantService{
-		tenantStore: storageMock,
-		log:         slog.Default(),
-	}
-	ctx := context.Background()
-	var t5s []*v1.Tenant
-	// filter by id
-	f1 := make(map[string]any)
-	tfr := &v1.TenantFindRequest{
-		Id: pointer.Pointer("t5"),
+func TestFindTenant(t *testing.T) {
+	ctx := t.Context()
+	ves := []datastore.Entity{
+		&v1.Project{},
+		&v1.ProjectMember{},
+		&v1.Tenant{},
+		&v1.TenantMember{},
 	}
 
-	f1["id"] = "t5"
-	storageMock.On("Find", ctx, f1, mock.AnythingOfType("*v1.Paging")).Return(t5s, nil, nil)
-	resp, err := ts.Find(ctx, tfr)
+	container, db, err := StartPostgres(ctx, ves...)
 	require.NoError(t, err)
-	assert.NotNil(t, resp)
-}
+	defer func() {
+		require.NoError(t, db.Close())
+		require.NoError(t, container.Terminate(ctx))
+	}()
 
-func TestFindTenantByName(t *testing.T) {
-	storageMock := &mocks.Storage[*v1.Tenant]{}
-	ts := &tenantService{
-		tenantStore: storageMock,
-		log:         slog.Default(),
+	var (
+		tenantStore = datastore.New(log, db, &v1.Tenant{})
+		testTenant1 = &v1.Tenant{
+			Meta: &v1.Meta{
+				Id:   "1",
+				Kind: "Tenant",
+
+				Apiversion: "v1",
+				Version:    1,
+				Annotations: map[string]string{
+					"a": "b",
+					"c": "d",
+				},
+				Labels: []string{"e", "f"},
+			},
+			Name:        "tenant-1",
+			Description: "tenant 1",
+		}
+		testTenant2 = &v1.Tenant{
+			Meta: &v1.Meta{
+				Id:         "2",
+				Kind:       "Tenant",
+				Apiversion: "v1",
+				Version:    1,
+				Annotations: map[string]string{
+					"c": "d",
+					"e": "f",
+				},
+				Labels: []string{"f", "g", "h"},
+			},
+			Name:        "tenant-2",
+			Description: "tenant 2",
+		}
+
+		service = &tenantService{
+			db:          db,
+			tenantStore: tenantStore,
+			log:         log,
+		}
+	)
+
+	tests := []struct {
+		name    string
+		prepare func()
+		req     *v1.TenantFindRequest
+		want    *v1.TenantListResponse
+		wantErr error
+	}{
+		{
+			name: "find by id",
+			req: &v1.TenantFindRequest{
+				Id: pointer.Pointer("1"),
+			},
+			prepare: func() {
+				require.NoError(t, tenantStore.Create(ctx, testTenant1))
+				require.NoError(t, tenantStore.Create(ctx, testTenant2))
+			},
+			want: &v1.TenantListResponse{
+				Tenants: []*v1.Tenant{
+					testTenant1,
+				},
+			},
+			wantErr: nil,
+		},
+		{
+			name: "find by id (no results)",
+			req: &v1.TenantFindRequest{
+				Id: pointer.Pointer("no-result"),
+			},
+			prepare: func() {
+				require.NoError(t, tenantStore.Create(ctx, testTenant1))
+				require.NoError(t, tenantStore.Create(ctx, testTenant2))
+			},
+			want: &v1.TenantListResponse{
+				Tenants: nil,
+			},
+			wantErr: nil,
+		},
+		{
+			name: "find by name",
+			req: &v1.TenantFindRequest{
+				Name: pointer.Pointer("tenant-2"),
+			},
+			prepare: func() {
+				require.NoError(t, tenantStore.Create(ctx, testTenant1))
+				require.NoError(t, tenantStore.Create(ctx, testTenant2))
+			},
+			want: &v1.TenantListResponse{
+				Tenants: []*v1.Tenant{
+					testTenant2,
+				},
+			},
+			wantErr: nil,
+		},
+		{
+			name: "find by annotation",
+			req: &v1.TenantFindRequest{
+				Annotations: map[string]string{
+					"a": "b",
+				},
+			},
+			prepare: func() {
+				require.NoError(t, tenantStore.Create(ctx, testTenant1))
+				require.NoError(t, tenantStore.Create(ctx, testTenant2))
+			},
+			want: &v1.TenantListResponse{
+				Tenants: []*v1.Tenant{
+					testTenant1,
+				},
+			},
+			wantErr: nil,
+		},
+		{
+			name: "find by annotation #2",
+			req: &v1.TenantFindRequest{
+				Annotations: map[string]string{
+					"a": "b",
+					"c": "d",
+				},
+			},
+			prepare: func() {
+				require.NoError(t, tenantStore.Create(ctx, testTenant1))
+				require.NoError(t, tenantStore.Create(ctx, testTenant2))
+			},
+			want: &v1.TenantListResponse{
+				Tenants: []*v1.Tenant{
+					testTenant1,
+				},
+			},
+			wantErr: nil,
+		},
+		{
+			name: "find by annotation #3",
+			req: &v1.TenantFindRequest{
+				Annotations: map[string]string{
+					"c": "d",
+				},
+			},
+			prepare: func() {
+				require.NoError(t, tenantStore.Create(ctx, testTenant1))
+				require.NoError(t, tenantStore.Create(ctx, testTenant2))
+			},
+			want: &v1.TenantListResponse{
+				Tenants: []*v1.Tenant{
+					testTenant1,
+					testTenant2,
+				},
+			},
+			wantErr: nil,
+		},
+		{
+			name: "find by label",
+			req: &v1.TenantFindRequest{
+				Labels: []string{"e"},
+			},
+			prepare: func() {
+				require.NoError(t, tenantStore.Create(ctx, testTenant1))
+				require.NoError(t, tenantStore.Create(ctx, testTenant2))
+			},
+			want: &v1.TenantListResponse{
+				Tenants: []*v1.Tenant{
+					testTenant1,
+				},
+			},
+			wantErr: nil,
+		},
+		{
+			name: "find by label #2",
+			req: &v1.TenantFindRequest{
+				Labels: []string{"e", "f"},
+			},
+			prepare: func() {
+				require.NoError(t, tenantStore.Create(ctx, testTenant1))
+				require.NoError(t, tenantStore.Create(ctx, testTenant2))
+			},
+			want: &v1.TenantListResponse{
+				Tenants: []*v1.Tenant{
+					testTenant1,
+				},
+			},
+			wantErr: nil,
+		},
+		{
+			name: "find by label #3",
+			req: &v1.TenantFindRequest{
+				Labels: []string{"f"},
+			},
+			prepare: func() {
+				require.NoError(t, tenantStore.Create(ctx, testTenant1))
+				require.NoError(t, tenantStore.Create(ctx, testTenant2))
+			},
+			want: &v1.TenantListResponse{
+				Tenants: []*v1.Tenant{
+					testTenant1,
+					testTenant2,
+				},
+			},
+			wantErr: nil,
+		},
 	}
-	ctx := context.Background()
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			for _, ve := range ves {
+				_, err := db.ExecContext(ctx, "TRUNCATE TABLE "+ve.TableName())
+				require.NoError(t, err)
+			}
 
-	// filter by name
-	var t6s []*v1.Tenant
-	tfr := &v1.TenantFindRequest{
-		Name: pointer.Pointer("Fifth"),
+			if tt.prepare != nil {
+				tt.prepare()
+			}
+
+			got, err := service.Find(ctx, tt.req)
+			if diff := cmp.Diff(err, tt.wantErr); diff != "" {
+				t.Errorf("(-want +got):\n%s", diff)
+				return
+			}
+			slices.SortFunc(got.Tenants, func(i, j *v1.Tenant) int {
+				if i.Meta.Id < j.Meta.Id {
+					return -1
+				} else {
+					return 1
+				}
+			})
+			if diff := cmp.Diff(tt.want, got, cmpopts.IgnoreTypes(protoimpl.MessageState{}), cmpopts.IgnoreFields(v1.Meta{}, "CreatedTime"), testcommon.IgnoreUnexported()); diff != "" {
+				t.Errorf("(-want +got):\n%s", diff)
+			}
+		})
 	}
-
-	f2 := make(map[string]any)
-	f2["tenant ->> 'name'"] = "Fifth"
-	storageMock.On("Find", ctx, f2, mock.AnythingOfType("*v1.Paging")).Return(t6s, nil, nil)
-	resp, err := ts.Find(ctx, tfr)
-	require.NoError(t, err)
-	assert.NotNil(t, resp)
 }
 
 func Test_tenantService_FindParticipatingProjects(t *testing.T) {
