@@ -1,13 +1,13 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.36.6
+// 	protoc-gen-go v1.36.8
 // 	protoc        (unknown)
 // source: v1/tenant.proto
 
 package v1
 
 import (
-	_type "github.com/metal-stack/masterdata-api/google/type"
+	postaladdress "google.golang.org/genproto/googleapis/type/postaladdress"
 	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
 	timestamppb "google.golang.org/protobuf/types/known/timestamppb"
@@ -774,6 +774,7 @@ type TenantFindRequest struct {
 	Annotations   map[string]string      `protobuf:"bytes,4,rep,name=annotations,proto3" json:"annotations,omitempty" protobuf_key:"bytes,1,opt,name=key" protobuf_val:"bytes,2,opt,name=value"`
 	Id            *string                `protobuf:"bytes,5,opt,name=id,proto3,oneof" json:"id,omitempty"`
 	Name          *string                `protobuf:"bytes,6,opt,name=name,proto3,oneof" json:"name,omitempty"`
+	Labels        []string               `protobuf:"bytes,7,rep,name=labels,proto3" json:"labels,omitempty"`
 	unknownFields protoimpl.UnknownFields
 	sizeCache     protoimpl.SizeCache
 }
@@ -834,6 +835,13 @@ func (x *TenantFindRequest) GetName() string {
 		return *x.Name
 	}
 	return ""
+}
+
+func (x *TenantFindRequest) GetLabels() []string {
+	if x != nil {
+		return x.Labels
+	}
+	return nil
 }
 
 type TenantResponse struct {
@@ -999,11 +1007,15 @@ func (x *PaymentDetails) GetLabels() []string {
 
 // ContactDetails contain more properties of a tenant
 type ContactDetails struct {
-	state       protoimpl.MessageState `protogen:"open.v1"`
-	Email       string                 `protobuf:"bytes,1,opt,name=email,proto3" json:"email,omitempty"`
-	PhoneNumber string                 `protobuf:"bytes,2,opt,name=phone_number,json=phoneNumber,proto3" json:"phone_number,omitempty"`
-	Address     *_type.PostalAddress   `protobuf:"bytes,3,opt,name=address,proto3" json:"address,omitempty"`
-	AvatarUrl   string                 `protobuf:"bytes,4,opt,name=avatar_url,json=avatarUrl,proto3" json:"avatar_url,omitempty"`
+	state protoimpl.MessageState `protogen:"open.v1"`
+	// Email of this tenant
+	Email string `protobuf:"bytes,1,opt,name=email,proto3" json:"email,omitempty"`
+	// Phone Number of this tenant
+	PhoneNumber string `protobuf:"bytes,2,opt,name=phone_number,json=phoneNumber,proto3" json:"phone_number,omitempty"`
+	// Address of this tenant
+	Address *postaladdress.PostalAddress `protobuf:"bytes,3,opt,name=address,proto3" json:"address,omitempty"`
+	// Avatar URL
+	AvatarUrl string `protobuf:"bytes,4,opt,name=avatar_url,json=avatarUrl,proto3" json:"avatar_url,omitempty"`
 	// Annotations of this payment details
 	Annotations map[string]string `protobuf:"bytes,5,rep,name=annotations,proto3" json:"annotations,omitempty" protobuf_key:"bytes,1,opt,name=key" protobuf_val:"bytes,2,opt,name=value"`
 	// Labels of this payment details
@@ -1056,7 +1068,7 @@ func (x *ContactDetails) GetPhoneNumber() string {
 	return ""
 }
 
-func (x *ContactDetails) GetAddress() *_type.PostalAddress {
+func (x *ContactDetails) GetAddress() *postaladdress.PostalAddress {
 	if x != nil {
 		return x.Address
 	}
@@ -1152,13 +1164,14 @@ const file_v1_tenant_proto_rawDesc = "" +
 	"\x02id\x18\x01 \x01(\tR\x02id\"U\n" +
 	"\x17TenantGetHistoryRequest\x12\x0e\n" +
 	"\x02id\x18\x01 \x01(\tR\x02id\x12*\n" +
-	"\x02at\x18\x02 \x01(\v2\x1a.google.protobuf.TimestampR\x02at\"\xab\x02\n" +
+	"\x02at\x18\x02 \x01(\v2\x1a.google.protobuf.TimestampR\x02at\"\xc3\x02\n" +
 	"\x11TenantFindRequest\x12\"\n" +
 	"\x06paging\x18\x03 \x01(\v2\n" +
 	".v1.PagingR\x06paging\x12H\n" +
 	"\vannotations\x18\x04 \x03(\v2&.v1.TenantFindRequest.AnnotationsEntryR\vannotations\x12\x13\n" +
 	"\x02id\x18\x05 \x01(\tH\x00R\x02id\x88\x01\x01\x12\x17\n" +
-	"\x04name\x18\x06 \x01(\tH\x01R\x04name\x88\x01\x01\x1a>\n" +
+	"\x04name\x18\x06 \x01(\tH\x01R\x04name\x88\x01\x01\x12\x16\n" +
+	"\x06labels\x18\a \x03(\tR\x06labels\x1a>\n" +
 	"\x10AnnotationsEntry\x12\x10\n" +
 	"\x03key\x18\x01 \x01(\tR\x03key\x12\x14\n" +
 	"\x05value\x18\x02 \x01(\tR\x05value:\x028\x01B\x05\n" +
@@ -1254,7 +1267,7 @@ var file_v1_tenant_proto_goTypes = []any{
 	(*IAMConfig)(nil),                         // 30: v1.IAMConfig
 	(*timestamppb.Timestamp)(nil),             // 31: google.protobuf.Timestamp
 	(*Paging)(nil),                            // 32: v1.Paging
-	(*_type.PostalAddress)(nil),               // 33: google.type.PostalAddress
+	(*postaladdress.PostalAddress)(nil),       // 33: google.type.PostalAddress
 }
 var file_v1_tenant_proto_depIdxs = []int32{
 	7,  // 0: v1.ListTenantMembersResponse.tenants:type_name -> v1.TenantWithMembershipAnnotations
