@@ -83,7 +83,7 @@ func (s *projectMemberService) Find(ctx context.Context, req *v1.ProjectMemberFi
 		f := fmt.Sprintf("projectmember -> 'meta' -> 'annotations' ->> '%s'", key)
 		filter[f] = value
 	}
-	res, _, err := s.projectMemberStore.Find(ctx, filter, nil)
+	res, _, err := s.projectMemberStore.Find(ctx, nil, filter)
 	if err != nil {
 		return nil, err
 	}
